+++ examples/game_example.py
@@ -2,7 +2,9 @@ import sys
 import os
 
 sys.path.insert(0, os.path.abspath(os.path.join(os.path.dirname(__file__), "..")))
-from project.game import Game, Bot
+
+from project.game_bot import Bot
+from project.game_play import Game
 
 
 def main():
