+++ project/game_bet.py
@@ -1,5 +1,5 @@
-from typing import List, Dict, Tuple, Optional, Protocol, Set
-from enum import Enum, auto
+from typing import List, Dict, Tuple, Optional
+from enum import Enum
 from abc import ABC, abstractmethod
 import random
 
@@ -382,250 +382,3 @@ class SnakeBet(Bet):
 
     def prints(self) -> str:
         return "SnakeBet"
-
-
-class Bot:
-    """Simulation of a player"""
-
-    def __init__(self, name: str, balance: int) -> None:
-        """
-        Initializes a bot.
-
-        Args:
-            name (str): Bot's name.
-            balance (int): Starting balance.
-        """
-        self.name = name
-        self.balance = balance
-        self.start_balance = balance
-        self.strategy: Optional[Strategy] = None
-
-    def choose_strategy(self) -> None:
-        """Selects a strategy based on balance."""
-        if self.balance < 100:
-            self.strategy = ClassicStrategy()
-        elif self.balance < 300:
-            self.strategy = RiskyStrategy()
-        elif self.balance < 500:
-            self.strategy = NoobStrategy()
-        elif self.balance < 700:
-            self.strategy = ProStrategy()
-        elif self.balance < 900:
-            self.strategy = ScamStrategy()
-        else:
-            self.strategy = AggressiveStrategy()
-
-    def make_bet(self, wheel: Wheel) -> Bet:
-        """
-        Makes a bet based on strategy.
-
-        Args:
-            wheel (Wheel): The game wheel.
-
-        Returns:
-            Bet: Chosen bet.
-        """
-        self.choose_strategy()
-        if self.strategy is None:
-            raise RuntimeError("Strategy is not set for this bot.")
-        return self.strategy.bet(self, wheel)
-
-    def update_balance(self, amount: int) -> None:
-        """
-        Updates bot's balance.
-
-        Args:
-            amount (int): The amount of points.
-        """
-        self.balance += amount
-
-    def get_strategy_name(self) -> str:
-        """Returns the name of strategy."""
-        if self.strategy:
-            return self.strategy.__class__.__name__
-        else:
-            return "None"
-
-
-class Strategy(ABC):
-    """Abstract class for strategies."""
-
-    @abstractmethod
-    def bet(self, bot: Bot, wheel: Wheel) -> Bet:
-        """Make a bet based on the bot's balance and strategy logic."""
-        pass
-
-
-class ClassicStrategy(Strategy):
-    """Classic strategy with colors bat."""
-
-    def bet(self, bot: Bot, wheel: Wheel) -> Bet:
-        """
-        Make a color bet.
-
-        Args:
-            bot (Bot): The bot makes the bet.
-            wheel (Wheel): The roulette wheel.
-
-        Returns:
-            ColorBet: color RED or BLACK.
-        """
-        amount = max(1, bot.balance // 10)
-        return ColorBet(random.choice([Color.RED, Color.BLACK]), amount)
-
-
-class RiskyStrategy(Strategy):
-    """Risky strategy with dozens bet."""
-
-    def bet(self, bot: Bot, wheel: Wheel) -> Bet:
-        """
-        Make a dozen bet.
-
-        Args:
-            bot (Bot): The bot makes the bet.
-            wheel (Wheel): The roulette wheel.
-
-        Returns:
-            DozenBet: Bet on random dozen.
-        """
-        amount = max(1, bot.balance // 5)
-        return DozenBet(random.choice([1, 2, 3]), amount)
-
-
-class NoobStrategy(Strategy):
-    """Classic strategy with snake bet."""
-
-    def bet(self, bot: Bot, wheel: Wheel) -> Bet:
-        """
-        Make a snake bet.
-
-        Args:
-            bot (Bot): The bot makes the bet.
-            wheel (Wheel): The roulette wheel.
-
-        Returns:
-            SnakeBet: Snake bet
-        """
-        amount = max(1, bot.balance // 20)
-        return SnakeBet(amount)
-
-
-class ProStrategy(Strategy):
-    """Pro strategy with street bet."""
-
-    def bet(self, bot: Bot, wheel: Wheel) -> Bet:
-        """
-        Make a street bet.
-
-        Args:
-            bot (Bot): The bot makes the bet.
-            wheel (Wheel): The roulette wheel.
-
-        Returns:
-            StreetBet: Bet on random street.
-        """
-        amount = max(1, bot.balance // 10)
-        return StreetBet(
-            random.choice([i for i in range(1, 35) if (i - 1) % 3 == 0]), amount
-        )
-
-
-class ScamStrategy(Strategy):
-    """Scam strategy with split bet."""
-
-    def bet(self, bot: Bot, wheel: Wheel) -> Bet:
-        """
-        Make a split bet.
-
-        Args:
-            bot (Bot): The bot makes the bet.
-            wheel (Wheel): The roulette wheel.
-
-        Returns:
-            SplitBet: Split bet.
-        """
-        a = random.randint(1, 35)
-        amount = max(1, bot.balance // 5)
-        return SplitBet((a, a + 1), amount)
-
-
-class AggressiveStrategy(Strategy):
-    """Aggressive strategy with single bet."""
-
-    def bet(self, bot: Bot, wheel: Wheel) -> Bet:
-        """
-        Make a single number bet.
-
-        Args:
-            bot (Bot): The bot makes the bet.
-            wheel (Wheel): The roulette wheel.
-
-        Returns:
-            SingleBet: Single number bet.
-        """
-        amount = max(1, bot.balance // 10)
-        return SingleBet(random.randint(1, 36), amount)
-
-
-class Game:
-    """Roulette game simulation."""
-
-    def __init__(self, bots: List[Bot], max_rounds: int) -> None:
-        """Initializes the game."""
-        self.bots = bots
-        self.max_rounds = max_rounds
-        self.wheel = Wheel()
-        self.round = 0
-
-    def play_round(self) -> None:
-        """Plays one round of the roulette game."""
-        self.round += 1
-        print(f"\nRound {self.round}")
-        number = self.wheel.spin()
-        color = self.wheel.get_color(number)
-        print(f"Winning number: {number}")
-        for bot in self.bots:
-            if bot.balance <= 0:
-                continue
-            print(f"\n{bot.name}")
-            print(f"Balance: {bot.balance}")
-            main_bet = bot.make_bet(self.wheel)
-            if main_bet.amount > bot.balance:
-                print("Not enough balance")
-                continue
-            win = 0
-            print(f"Bet amount: {main_bet.amount}")
-            if main_bet.is_win(number, self.wheel):
-                payout = main_bet.amount * main_bet.coefficient()
-                print(f"Bet: {main_bet.prints()}")
-                print(f"Coefficient: {main_bet.coefficient()}")
-                print(f"Result: Wins {payout}")
-                win += payout
-                bot.update_balance(win)
-            else:
-                print(f"Bet: {main_bet.prints()}")
-                print(f"Result: Loses {main_bet.amount}")
-                bot.update_balance(-main_bet.amount)
-
-            print(f"Strategy: {bot.get_strategy_name()}")
-            print(f"New balance: {bot.balance}")
-
-    def play(self) -> None:
-        """Simulate all rounds of the game."""
-        for i in range(self.max_rounds):
-            self.play_round()
-        print("\nGame Over")
-        bot_profits = []
-        for bot in self.bots:
-            profit = bot.balance - bot.start_balance
-            bot_profits.append((bot, profit))
-            print(f"{bot.name}: Final Balance = {bot.balance}")
-            if profit > 0:
-                result = "Wins"
-                print(f"{result} {profit}")
-            elif profit < 0:
-                result = "Loses"
-                print(f"{result} {profit}")
-            else:
-                result = "Same"
-                print(f"{result} {profit}")
